refactor(src): consolidate sync cleanup, split fsops, drop dead code
* sync: collapse three near-identical remove_stale_* functions
  (skills, commands, mcps) into a single generic helper in sync/mod.rs
  taking a per-kind teardown closure. Bookkeeping (summary bump, Action
  push, dry-run branching) now lives in one place.
* sync: introduce SyncMut<'a> struct bundling (state, runtime, summary,
  actions) and thread it through skills.rs helpers. Removes three
  #[allow(clippy::too_many_arguments)] attributes.
* sync: drop unused _ctx parameter from needs_fetch_commands.
* fsops: split mod.rs (854 → 618 lines) by extracting config loading
  and the recursive `extends:` merge into fsops/config.rs (254 lines).
  Re-export load_config_any so external call sites are unchanged.
* self_update: delete the dead _placeholder() function and a stray
  `let _ = ACCENT;` warning-suppression line.
* mcps/merge: remove unused `_name` parameter from `identity()`;
  update its single call site to wrap with a closure.
